requirements.txt
amqp==5.0.9
api==0.0.7
asgiref==3.4.1
automium==0.2.6
automium-web==0.1.1
backports.zoneinfo==0.2.1
billiard==3.6.4.0
celery==5.2.3
certifi==2021.10.8
charset-normalizer==2.0.10
click==8.0.3
click-aliases==1.0.1
click-default-group==1.2.2
click-didyoumean==0.3.0
click-help-colors==0.9.1
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
commonmark==0.9.1
Deprecated==1.2.13
Django==4.0.1
djangorestframework==3.13.1
Flask==2.0.2
greenlet==1.1.2
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
kombu==5.2.3
legacy==0.1.6
MarkupSafe==2.0.1
nose==1.3.7
packaging==21.3
prompt-toolkit==3.0.24
Pygments==2.11.2
pyparsing==3.0.6
python-dateutil==2.8.2
pytz==2021.3
redis==4.1.0
requests==2.27.1
rich==9.13.0
six==1.16.0
SQLAlchemy==1.4.29
sqlparse==0.4.2
tabulate==0.8.9
task==0.2.5
typing-extensions==3.10.0.2
urllib3==1.26.8
vine==5.0.0
wcwidth==0.2.5
Werkzeug==2.0.2
wrapt==1.13.3
